Created new project: XRibTiles
The new project generalizes programs for RibTile project.
The new project is meant to work for arbitrary n (number of squares
in a ribbon tile) and for a rather general family of regions. Namely,
it is supposed to work for regions that have only one non-trivial
intersection with each of the level lines x + y = l.