Significant update. The aim is to tile an arbitrary simply-conn. region
The most important new class is the Interval Graph, which 
encodes the region as the system of intervals which are intersections of
the region with level lines x + y = l. 
The intervals can be connected reflecting the topology of
the region. For simply-connected regions this graph is a
tree and the tiling will be obtained by processing vertices in this
tree. 

Signed-off-by: slavakargin <slavakargin@yahoo.com>